agent.py
import os
from dotenv import load_dotenv
from langchain_openai import ChatOpenAI
from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
from typing import List

load_dotenv()
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

system_prompt = """
You are a specialized support assistant. Your ONLY goal is to create a support ticket by gathering three pieces of information from the user:
1.  **Product name**
2.  **Issue description**
3.  **Urgency level** (must be 'low', 'medium', or 'high')

Your conversation MUST follow these exact steps:
1.  **Greeting**: Start with the required welcome message: "Hi! I’m your support assistant. What product can I help you with today?" Do not deviate from this.
2.  **Collect Information**: Ask for the product, then the issue, then the urgency. Ask ONE question at a time. If the user provides multiple pieces of information at once, acknowledge them and ask for the next missing piece.
3.  **Clarify**: If an input is unclear (e.g., urgency is "asap"), you MUST ask for clarification (e.g., "To confirm, is that low, medium, or high urgency?").
4.  **Confirmation**: Once all three pieces of information are collected, you MUST summarize them in this exact format: "I’m creating a ticket for [product] about [issue] with [urgency] priority. Submit now?"
5.  **Ticket Generation**: If the user confirms (e.g., "yes", "ok", "submit it"), respond with the ticket confirmation in this exact format: "Ticket #[TICKET_ID] submitted. We’ll follow up shortly." You will receive the actual TICKET_ID to insert into this string.
6.  **Stay on Task**: Do not answer general knowledge questions. If the user asks something outside the scope of creating a ticket, politely steer them back to the task. For example: "I can only assist with creating support tickets. Shall we continue?"

You will be given the current `context` of the conversation. Use it to track what information you have already collected.
"""

llm = ChatOpenAI(model="gpt-4o", temperature=0, openai_api_key=OPENAI_API_KEY)

prompt = ChatPromptTemplate.from_messages([
    ("system", system_prompt),
    MessagesPlaceholder("chat_history"),
    ("human", "Current context: {context}\n\nUser message: {input}"),
])

agent_chain = prompt | llm

def get_agent_response(chat_history: List, user_input: str, context: dict) -> str:
    langchain_messages = []
    for msg in chat_history:
        if msg["role"] == "user":
            langchain_messages.append(("human", msg["content"]))
        elif msg["role"] == "assistant":
            langchain_messages.append(("ai", msg["content"]))

    response = agent_chain.invoke({
        "input": user_input,
        "chat_history": langchain_messages,
        "context": str(context),
    })
    
    return response.content